셀레니움 추가 레진 파싱 진행중 레진 ID값이 상당히 큰 관계로 WebtoonDTO의 titleId int-long으로 변경

+++ toonalyst/src/main/java/com/toonalyst/domain/webtoon/CategoryDTO.java
@@ -23,6 +23,6 @@ import lombok.ToString;
 @Setter
 @ToString
 public class CategoryDTO {
-	private int titleId;
+	private long titleId;
 	private String category;
 }
