+++ toonalyst/src/main/java/com/toonalyst/persistence/webtoon/WebtoonDAO.java
@@ -5,5 +5,5 @@ import com.toonalyst.domain.webtoon.WebtoonDTO;
 public interface WebtoonDAO {
 
 	public void create(WebtoonDTO wDto);
-	public WebtoonDTO selectone(int titleId);
+	public WebtoonDTO selectone(long titleId);
 }
