+++ toonalyst/src/main/java/com/toonalyst/service/webtoon/WebtoonService.java
@@ -4,5 +4,5 @@ import com.toonalyst.domain.webtoon.WebtoonDTO;
 
 public interface WebtoonService {
 	public void create(WebtoonDTO wDto);
-	public WebtoonDTO selectone(int titleId);
+	public WebtoonDTO selectone(long titleId);
 }
